requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.7
colorama==0.4.6
crispy-bootstrap5==2026.3
dj-database-url==3.1.2
dj-rest-auth==7.2.0
Django==6.0.3
django-allauth==65.15.0
django-cors-headers==4.9.0
django-crispy-forms==2.6
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
